vendor_management/requirements.txt
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
virtualenv==20.25.0